Fix movie decoding on x64
Closes #107

+++ src/movie_lib.cc
@@ -1865,7 +1865,7 @@ static void _nfPkDecomp(unsigned char* a1, unsigned char* a2, int a3, int a4, in
     int v7;
     int i;
     int j;
-    int v10;
+    ptrdiff_t v10;
     int v11;
     int v13;
     int byte;
